requirements.txt
matplotlib==3.10.6
panda==0.3.1
pip==24.0
plotly==6.3.0
yfinance==0.2.65
